+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# hpqtypes-effectful-1.0.1.0 (????-??-??)
+* Compatibility with `hpqtypes` >= 1.12.0.0.
+
 # hpqtypes-effectful-1.0.0.1 (2023-03-28)
 * Compatibility with `hpqtypes` >= 1.11.1.0.